Fix compilation errors in CLI modules
- edit.rs: Use ref binding to avoid partial move
- embed.rs: Use model field instead of model_name
- doctor.rs: Use checks field instead of results
- interactive.rs: Cast u64 to usize for commands_executed

+++ src/cli/interactive.rs
@@ -106,7 +106,7 @@ pub async fn run(
 
     // Emit interactive mode ended event
     let duration_ms = start_time.elapsed().as_millis() as u64;
-    let total_commands = commands_executed.load(Ordering::Relaxed);
+    let total_commands = commands_executed.load(Ordering::Relaxed) as usize;
     event_helpers::emit_domain_event(AgentEvent::InteractiveModeEnded {
         commands_executed: total_commands,
         duration_ms,
